created a fetch-folder in the utils folder and moved fetch-files there
